requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0
PyMySQL==1.1.1
python-dotenv==1.0.1
google-api-python-client==2.145.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
twilio==9.3.7
openai==1.51.2
python-multipart==0.0.12
PyMuPDF==1.23.8
numpy==2.3.3

